modifications for C90

+++ arch/i386/kernel/types.h
@@ -2,16 +2,19 @@
 
 #ifndef __types_h__
 #define __types_h__
+#include <stdint.h>
 
 typedef unsigned char u8_t;			/* unsigned types	*/
 typedef unsigned short int u16_t;
 typedef unsigned long int u32_t;
-typedef unsigned long long int u64_t;
+/* // typedef unsigned long long int u64_t; */
+typedef uint64_t u64_t;
 
 typedef signed char s8_t;			/* signed types		*/
 typedef signed short int s16_t;
 typedef signed long int s32_t;
-typedef signed long long int s64_t;
+/* // typedef signed long long int s64_t; */
+typedef int64_t s64_t;
 
 #define NULL ((void *)0)
